+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **`[spinner]`** Added support of different duration formats
 - **`[strutil]`** Added method `Wrap`
 - **`[cron]`** Code refactoring
+- **`[options]`** Code refactoring
 - **`[mathutil]`** Methods `Min` and `Max` marked as deprecated
 - **`[fmtc]`** Deprecated methods have been removed
 - **`[fsutil]`** Deprecated methods have been removed